merging polyglot crate from polyglote branch
I just kept the most important things to generate the types.
I would also like to extract the visualization into another crate. This way it also helps with finding a good interface to consume the ast node types.

The generation part will still need some work to make it more usable and automatically generate the type files. Next step would be to cleanup and reorganize the main features to make it more resilient. At some point I want to use the queries to do the generation