adding roles aka fields to nodes of the hyperast
will need to investigate if it is derived data or primary,
probably derived because the parser cannot choose between 2 rules that only differ by their fields,
nonetheless is might be possible in case of (improper?) uses of aliases,
ie. aliases with the name of another rule should produce a subset of the other rule.