making TypeStore a marker, i.e., PhantomType
there is no point holding non-compile time stuff there. If really nessessary, lets use some static var. The big benefit of the marker is for ease of use of the HyperAST, and to enable easy/safe transmute.
Other features that could be handled specifically for types, can probably be provided generally to all metadata, so lets handle that in the node_store or on the hyperast.