Fix hybrid mapper benchmark and lazification
The benchmarks were not using the lazy subtree for both variants,
combined with only measuring the bottomup phase, it was hidding part of
the mapping cost.

I also fixed some inefficiencies in the use of lazyfication for the
recovery phase of hybrid, mainly related to not having used
decompress_children.