setup cs to use the same kind of leaf iterator
Doing it on both reduces the confusion variables It is still in line
with the original description, as we are still iterating leafs in
post-order. But instead we traverse in pre-order and stop descending
when meeting a leaf. It enables lazy decompression related
optimizations, we won't descompress further than statements.

Interpreting the orginal post-order description, the one implementing
the is_leaf predicate should be cautious and not return true if the same
is_leaf would return true on a descendant ! Obviously, to avoid loosing
the benefits of the lazy decompression, it requires to precompute and
diffuse the verdict of the predicate on subtrees